requirements.txt
# RAG / LangChain
langchain>=1.1.3
langchain-community>=0.4.1
langchain-mistralai>=1.1.0
faiss-cpu>=1.13.1
mistralai>=1.9.11

# Data
pandas>=2.3.3
python-dateutil>=2.9.0.post0

# Environment / API
python-dotenv>=1.2.1
requests>=2.32.5

# UI / Streamlit
streamlit>=1.52.1

# Tests unitaires
pytest>=9.0.2
